+++ CHANGELOG.md
@@ -2,6 +2,7 @@ History of user-visible changes.
 
 ## [v1.6.3](https://gitlab.com/openconnect/openconnect-gui/-/tree/main) (unreleased)
 - Limit the auto-generated and user-supplied interface name to the maximum length allowed by libopenconnect (#365, #378)
+- Store username and password for GlobalProtect profiles (#390)
 
 [Full Changelog](https://gitlab.com/openconnect/openconnect-gui/-/compare/v1.6.2...main)